[arcane,accelerator] Renomme 'Runner::Impl' en 'RunnerImpl' et déplace sa définition dans son propre fichier.
